sprint 14 v1

+++ controllers/cards.js
@@ -28,15 +28,23 @@ module.exports.createCard = (req, res) => {
 };
 
 module.exports.deleteCard = (req, res) => {
-  Card.findByIdAndRemove(req.params.cardId)
-    .then((card) => {
-      if (card) {
-        res.status(204).send();
-      } else {
-        res.status(404).json({ message: 'Карточка не найдена' });
+  Card.findById(req.params.cardId)
+    .then(res=>{
+      if (req.user._id===res[0].owner){
+        Card.findByIdAndRemove(req.params.cardId)
+          .then((card) => {
+            if (card) {
+              res.status(204).send();
+            } else {
+              res.status(404).json({ message: 'Карточка не найдена' });
+            }
+          })
+          .catch(() => {
+            res.status(500).json({ message: 'Произошла ошибка' });
+          });
       }
     })
-    .catch(() => {
-      res.status(500).json({ message: 'Произошла ошибка' });
-    });
+    .catch(err=>{
+      res.status(403).end();
+    })
 };
